// src/app/modules/academics/student-admissions/model/student-emergency-contact-details-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Gets or Sets StudentEmergencyContactView
 */
export interface StudentEmergencyContactDetailsView { 
    /**
     * Gets or sets EmergencyContactName1
     */
    emergencyContactName1?: string;
    /**
     * Gets or sets EmergencyContactRelation1
     */
    emergencyContactRelation1?: string;
    /**
     * Gets or sets EmgergencyContactTelNumber1
     */
    emgergencyContactTelNumber1?: string;
    /**
     * Gets or sets EmergencyContactName2
     */
    emergencyContactName2?: string;
    /**
     * Gets or sets EmergencyContactRelation2
     */
    emergencyContactRelation2?: string;
    /**
     * Gets or sets EmgergencyContactTelNumber2
     */
    emgergencyContactTelNumber2?: string;
    /**
     * Gets or sets NumberOfOlderSibling
     */
    numberOfOlderSibling?: number;
    /**
     * Gets or sets NumberOfYoungerSibling
     */
    numberOfYoungerSibling?: number;
    /**
     * Gets or sets SocialCategory
     */
    socialCategory?: string;
}