// src/app/modules/academics/student-admissions/model/studentBulkPostView.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { StudentDetailsViewModel } from './student-details-view-model';
import { AttachmentViewModel } from './attachment-view-model';


/**
 * StudentBulkPostViewResult
 */
export interface StudentBulkPostView { 
    /**
     * StudentDetails ViewModel
     */
    studentDetailsViewModel?: Array<StudentDetailsViewModel>;
    /**
     * Attachment ViewModel
     */
    attachmentViewModel?: AttachmentViewModel;
}