// src/app/modules/configuration/class-schedules/model/holiday-bulk-post-view.ts
import { HolidaysView } from './holidays-view';
import { AttachmentViewModel } from 'app/modules/academics/student-information-system/model/attachment-view-model';

/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



/**
 * HolidayBulkPostView
 */
export interface HolidayBulkPostView { 
    /**
     * Holiday ViewModels
     */
    holidayViewModels?: Array<HolidaysView>;
    /**
     * Attachment ViewModel
     */
    attachmentViewModel?: AttachmentViewModel;
}