// src/app/modules/staff/models/staff-bulk-post-view.ts
import { StaffView } from 'app/models/staff/staff-view';
import { AttachmentViewModel } from '../staff-admissions/model/attachment-view-model';

/**
 * Teacher API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StaffBulkPostView
 */
export interface StaffBulkPostView { 
    /**
     * StaffView
     */
    staffView?: Array<StaffView>;
    /**
     * Attachment ViewModel
     */
    attachmentViewModel?: AttachmentViewModel;
}